partial commit

+++ dws-jpa/src/main/java/es/cesguiro/domain/usecase/book/BookCountUseCase.java
@@ -2,5 +2,5 @@ package es.cesguiro.domain.usecase.book;
 
 public interface BookCountUseCase {
 
-    int execute();
+    long execute();
 }
